feat: move NIP-86 relay management infrastructure to commons
- Create INip86Retriever interface in commons commonMain for KMP
- Move Nip86Retriever (OkHttp impl) from amethyst to commons jvmAndroid
- Create RelayManagementState in commons commonMain with all NIP-86
  relay management state and operations (platform-agnostic)
- Refactor RelayManagementViewModel to delegate to RelayManagementState
- Add okhttp/okhttpCoroutines deps to commons jvmAndroid source set

The Android ViewModel retains LocalCache-dependent PubkeyUser resolution
while all core state management is now shareable with iOS.

Part of the KMP iOS migration tracked in #2238.
