+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/model/nip86RelayManagement/RelayManagementState.kt
@@ -1,0 +1,327 @@
+/*
+ * Copyright (c) 2025 Vitor Pamplona
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+ * Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
+ * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package com.vitorpamplona.amethyst.commons.model.nip86RelayManagement
+
+import androidx.compose.runtime.Stable
+import com.vitorpamplona.quartz.nip86RelayManagement.Nip86Client
+import com.vitorpamplona.quartz.nip86RelayManagement.rpc.AllowedPubkey
+import com.vitorpamplona.quartz.nip86RelayManagement.rpc.BannedEvent
+import com.vitorpamplona.quartz.nip86RelayManagement.rpc.BannedPubkey
+import com.vitorpamplona.quartz.nip86RelayManagement.rpc.BlockedIp
+import com.vitorpamplona.quartz.nip86RelayManagement.rpc.EventNeedingModeration
+import com.vitorpamplona.quartz.nip86RelayManagement.rpc.Nip86Request
+import kotlinx.collections.immutable.ImmutableList
+import kotlinx.collections.immutable.persistentListOf
+import kotlinx.collections.immutable.toImmutableList
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.launch
+
+/**
+ * Platform-agnostic state holder for NIP-86 relay management.
+ *
+ * Contains all NIP-86 relay management state and operations, decoupled from
+ * Android ViewModel. Can be used from any CoroutineScope (ViewModel, iOS, etc.)
+ *
+ * @param client The NIP-86 RPC client for a specific relay
+ * @param retriever Platform-specific HTTP implementation
+ * @param scope Coroutine scope for launching async operations
+ */
+@Stable
+class RelayManagementState(
+    val client: Nip86Client,
+    private val retriever: INip86Retriever,
+    private val scope: CoroutineScope,
+) {
+    private val _supportedMethods = MutableStateFlow<ImmutableList<String>>(persistentListOf())
+    val supportedMethods: StateFlow<ImmutableList<String>> = _supportedMethods
+
+    private val _bannedPubkeys = MutableStateFlow<List<BannedPubkey>>(emptyList())
+    val bannedPubkeys: StateFlow<List<BannedPubkey>> = _bannedPubkeys
+
+    private val _allowedPubkeys = MutableStateFlow<List<AllowedPubkey>>(emptyList())
+    val allowedPubkeys: StateFlow<List<AllowedPubkey>> = _allowedPubkeys
+
+    private val _bannedEvents = MutableStateFlow<List<BannedEvent>>(emptyList())
+    val bannedEvents: StateFlow<List<BannedEvent>> = _bannedEvents
+
+    private val _eventsNeedingModeration = MutableStateFlow<List<EventNeedingModeration>>(emptyList())
+    val eventsNeedingModeration: StateFlow<List<EventNeedingModeration>> = _eventsNeedingModeration
+
+    private val _allowedKinds = MutableStateFlow<List<Int>>(emptyList())
+    val allowedKinds: StateFlow<List<Int>> = _allowedKinds
+
+    private val _blockedIps = MutableStateFlow<List<BlockedIp>>(emptyList())
+    val blockedIps: StateFlow<List<BlockedIp>> = _blockedIps
+
+    private val _isLoading = MutableStateFlow(false)
+    val isLoading: StateFlow<Boolean> = _isLoading
+
+    private val _error = MutableStateFlow<String?>(null)
+    val error: StateFlow<String?> = _error
+
+    fun loadSupportedMethods() {
+        scope.launch {
+            _isLoading.value = true
+            _error.value = null
+            val response = retriever.execute(client, Nip86Request.supportedMethods())
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                _supportedMethods.value = client.parseSupportedMethods(response)?.toImmutableList() ?: persistentListOf()
+            }
+            _isLoading.value = false
+        }
+    }
+
+    fun loadBannedPubkeys() {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.listBannedPubkeys())
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                _bannedPubkeys.value = client.parseBannedPubkeys(response)?.distinctBy { it.pubkey } ?: emptyList()
+            }
+        }
+    }
+
+    fun loadAllowedPubkeys() {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.listAllowedPubkeys())
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                _allowedPubkeys.value = client.parseAllowedPubkeys(response)?.distinctBy { it.pubkey } ?: emptyList()
+            }
+        }
+    }
+
+    fun loadBannedEvents() {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.listBannedEvents())
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                _bannedEvents.value = client.parseBannedEvents(response)?.distinctBy { it.id } ?: emptyList()
+            }
+        }
+    }
+
+    fun loadEventsNeedingModeration() {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.listEventsNeedingModeration())
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                _eventsNeedingModeration.value = client.parseEventsNeedingModeration(response)?.distinctBy { it.id } ?: emptyList()
+            }
+        }
+    }
+
+    fun loadAllowedKinds() {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.listAllowedKinds())
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                _allowedKinds.value = client.parseAllowedKinds(response)?.distinctBy { it } ?: emptyList()
+            }
+        }
+    }
+
+    fun loadBlockedIps() {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.listBlockedIps())
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                _blockedIps.value = client.parseBlockedIps(response)?.distinctBy { it.ip } ?: emptyList()
+            }
+        }
+    }
+
+    fun banPubkey(
+        pubkey: String,
+        reason: String? = null,
+    ) {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.banPubkey(pubkey, reason))
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                loadBannedPubkeys()
+            }
+        }
+    }
+
+    fun unbanPubkey(pubkey: String) {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.unbanPubkey(pubkey))
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                loadBannedPubkeys()
+            }
+        }
+    }
+
+    fun allowPubkey(
+        pubkey: String,
+        reason: String? = null,
+    ) {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.allowPubkey(pubkey, reason))
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                loadAllowedPubkeys()
+            }
+        }
+    }
+
+    fun unallowPubkey(pubkey: String) {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.unallowPubkey(pubkey))
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                loadAllowedPubkeys()
+            }
+        }
+    }
+
+    fun banEvent(
+        eventId: String,
+        reason: String? = null,
+    ) {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.banEvent(eventId, reason))
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                loadBannedEvents()
+            }
+        }
+    }
+
+    fun allowEvent(
+        eventId: String,
+        reason: String? = null,
+    ) {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.allowEvent(eventId, reason))
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                loadEventsNeedingModeration()
+            }
+        }
+    }
+
+    fun changeRelayName(newName: String) {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.changeRelayName(newName))
+            if (response.error != null) {
+                _error.value = response.error
+            }
+        }
+    }
+
+    fun changeRelayDescription(newDescription: String) {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.changeRelayDescription(newDescription))
+            if (response.error != null) {
+                _error.value = response.error
+            }
+        }
+    }
+
+    fun changeRelayIcon(newIconUrl: String) {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.changeRelayIcon(newIconUrl))
+            if (response.error != null) {
+                _error.value = response.error
+            }
+        }
+    }
+
+    fun allowKind(kind: Int) {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.allowKind(kind))
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                loadAllowedKinds()
+            }
+        }
+    }
+
+    fun disallowKind(kind: Int) {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.disallowKind(kind))
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                loadAllowedKinds()
+            }
+        }
+    }
+
+    fun blockIp(
+        ip: String,
+        reason: String? = null,
+    ) {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.blockIp(ip, reason))
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                loadBlockedIps()
+            }
+        }
+    }
+
+    fun unblockIp(ip: String) {
+        scope.launch {
+            val response = retriever.execute(client, Nip86Request.unblockIp(ip))
+            if (response.error != null) {
+                _error.value = response.error
+            } else {
+                loadBlockedIps()
+            }
+        }
+    }
+
+    fun clearError() {
+        _error.value = null
+    }
+
+    fun loadAllLists() {
+        val methods = _supportedMethods.value
+        if (methods.contains("listbannedpubkeys")) loadBannedPubkeys()
+        if (methods.contains("listallowedpubkeys")) loadAllowedPubkeys()
+        if (methods.contains("listbannedevents")) loadBannedEvents()
+        if (methods.contains("listeventsneedingmoderation")) loadEventsNeedingModeration()
+        if (methods.contains("listallowedkinds")) loadAllowedKinds()
+        if (methods.contains("listblockedips")) loadBlockedIps()
+    }
+}
